vtkQtLabelRenderStrategyInternals: expose in a header
The implementation detail class `vtkQtLabelRenderStrategy::Internals` is
moved to an own header file `vtkQtLabelRenderStrategyInternals.h` as
otherwise it is not possible to cleanly derive from
`vtkQtLabelRenderStrategy`: a derived class needs access to the
`Internals` to be able to modify relevant functionality and therefore it
must see the definition of that class.

Fixes #17395
